Importing: seiscomp3-zurich-2013.043.03-src.tar.gz

+++ src/core/libs/seiscomp3/communication/connectioninfo.cpp
@@ -248,7 +248,7 @@ std::string ConnectionInfo::info(const NetworkInterface* ni)
 
 	os << ConnectionInfoTag(TOTAL_MEMORY_TAG).toString() << "=" << _totalMemory << "&";
 
-	double currentMemoryUsage = memoryUsage();
+	int currentMemoryUsage = memoryUsage();
 	os << ConnectionInfoTag(CLIENT_MEMORY_USAGE_TAG).toString() << "=" << currentMemoryUsage << "&";
 
 	return os.str();
